Remove references to command-specific options.
As far as I can tell, there are no commands that have options specific
to that command (as opposed to global options). Also, we're using the
stdlib flag package in such a way that we can't distinguish anyway. So
for now, let's get rid of references to command-specific options.

doozer [OPTIONS] CMD [OPTIONS] ARGS

becomes

doozer [OPTIONS] CMD ARGS